1. Moved from Maven to SBT 2. Moved from jUnit to ScalaTest 3. Moved to Scala 2.11 4. Minor code clean-ups.
