+++ src/main/scala/hammurabi/RuleEngine.scala
@@ -17,32 +17,29 @@ class RuleEngine(rules: Traversable[Rule]) {
 
   def execOn(workingMemory: WorkingMemory): Option[Any] = {
     this.workingMemory = workingMemory
-    evaluate(rules.map(new RuleEvaluator(_, workingMemory)).par) fold (
-      t => throw t
-    , res => res
-    )
+    evaluate(rules.map(new RuleEvaluator(_, workingMemory)).par) fold (t => throw t, res => res)
   }
 
   @tailrec
   private def evaluate(evaluators: ParIterable[RuleEvaluator]): Either[Throwable, Option[Any]] = {
-    sequence(evaluators.map(_.evaluate)) match {
+    sequence(evaluators.map(_.evaluate())) match {
       case Left(Nil) => throw new Exception("We should not be here!")
       case Left(e :: es) => Left(e)
       case Right(Nil) => Right(None)
-      case Right(l) => {
-        val execRes = l sortWith(_.salience > _.salience) map (_.execRule) filter (_.isDefined)
-        if (!execRes.isEmpty) Right(execRes.head) else evaluate(evaluators)
-      }
+      case Right(l) =>
+        val execRes = l sortWith(_.salience > _.salience) map (_.execRule()) filter (_.isDefined)
+        if (execRes.nonEmpty)
+          Right(execRes.head) else evaluate(evaluators)
     }
   }
 
   private def sequence[A](t: GenIterable[Either[Throwable, List[A]]]): Either[List[Throwable], List[A]] = {
     var items: List[A] = Nil
     var errors: List[Throwable] = Nil
-    t foreach { _ match {
+    t foreach {
       case Left(ex) => errors = ex :: errors
       case Right(l) => items = items ::: l
-    }}
+    }
     if (errors.isEmpty) Right(items) else Left(errors)
   }
 }
