+++ src/main/scala/hammurabi/RuleEvaluator.scala
@@ -23,6 +23,7 @@ private[hammurabi] abstract class RuleManipulator(workingMemory: WorkingMemory) 
   def anyHaving[A](clazz: Class[A])(condition: A => Boolean): Option[A] = fetch(clazz)(workingMemory.allHaving(clazz)(condition))
 
   def +(item: Any) = workingMemory + item
+
   def -(item: Any) = workingMemory - item
 
   def exitWith(result: Any)
@@ -31,7 +32,7 @@ private[hammurabi] abstract class RuleManipulator(workingMemory: WorkingMemory) 
 }
 
 private[hammurabi] class RuleEvaluator(rule: Rule, workingMemory: WorkingMemory)
-                  extends RuleManipulator(workingMemory) with Logger {
+  extends RuleManipulator(workingMemory) with Logger {
 
   var executedSets = Set[RuleExecutionSet]()
   var currentExecutionSet: RuleExecutionSet = _
@@ -42,7 +43,7 @@ private[hammurabi] class RuleEvaluator(rule: Rule, workingMemory: WorkingMemory)
   def evaluate(): Either[Throwable, List[RuleExecutor]] = {
     allCatch[List[RuleExecutor]] either {
       var executors: List[RuleExecutor] = List[RuleExecutor]()
-      initFirstExecution
+      initFirstExecution()
       executors = executors +? evalRule
       isFirstExection = false
       while (valuesCombinator.hasNext) executors = executors +? evalRule
@@ -50,7 +51,7 @@ private[hammurabi] class RuleEvaluator(rule: Rule, workingMemory: WorkingMemory)
     }
   }
 
-  private def initFirstExecution = {
+  private def initFirstExecution() = {
     isFirstExection = true
     valuesCombinator = new ValuesCombinator
   }
@@ -78,22 +79,22 @@ private[hammurabi] class RuleEvaluator(rule: Rule, workingMemory: WorkingMemory)
     executedSets = executedSets + executedSet
   }
 
-  private def isRuleFinished = (!isFirstExection && !valuesCombinator.hasNext)
+  private def isRuleFinished = !isFirstExection && !valuesCombinator.hasNext
 
   protected def fetch[A](clazz: Class[A])(f: => List[A]): Option[A] = {
     currentExecutionSet += (
       if (isFirstExection) valuesCombinator += f.asInstanceOf[Traversable[A]]
       else valuesCombinator.next(clazz)
-    )
+      )
   }
 }
 
 private[hammurabi] class RuleExecutor(evaluator: RuleEvaluator, rule: Rule, workingMemory: WorkingMemory, executionSet: RuleExecutionSet)
-                  extends RuleManipulator(workingMemory) with Logger {
+  extends RuleManipulator(workingMemory) with Logger {
 
   val salience = rule.salience
 
-  private val executionIterator = executionSet.toIterator
+  private val executionIterator = executionSet.toIterator()
   private var result: Option[Any] = None
 
   def execRule(): Option[Any] = {
@@ -109,7 +110,7 @@ private[hammurabi] class RuleExecutor(evaluator: RuleEvaluator, rule: Rule, work
   }
 
   protected def fetch[A](clazz: Class[A])(f: => List[A]): Option[A] =
-    executionIterator.next.asInstanceOf[Option[A]]
+    executionIterator.next().asInstanceOf[Option[A]]
 
   override def exitWith(result: Any) = this.result = Some(result)
 }
@@ -117,21 +118,22 @@ private[hammurabi] class RuleExecutor(evaluator: RuleEvaluator, rule: Rule, work
 private class RuleExecutionSet {
   val executionSet = new mutable.ListBuffer[Option[_]]()
 
-  def +=[A](item: Option[A]) = { executionSet += item; item }
+  def +=[A](item: Option[A]) = {
+    executionSet += item; item
+  }
 
   def toIterator(): Iterator[Option[_]] = executionSet.iterator
 
   override def equals(that: Any): Boolean = that match {
-    case that: RuleExecutionSet => {
+    case that: RuleExecutionSet =>
       if (this.executionSet.length == that.executionSet.length) {
         val thatIterator = that.executionSet.toIterator
-        (true /: executionSet) (_ && _ == thatIterator.next)
+        (true /: executionSet)(_ && _ == thatIterator.next)
       } else false
-    }
     case _ => false
   }
 
-  override def hashCode = (0 /: executionSet) (_ + _.hashCode * 13)
+  override def hashCode = (0 /: executionSet)(_ + _.hashCode * 13)
 
   override def toString = "[" + executionSet.map(_.getOrElse("Nothing")).mkString(", ") + "]"
 
@@ -145,9 +147,9 @@ private class ValuesCombinator {
   lazy val valuesIterator = cartesianProduct(values).tail.flatten.toIterator
   var finished = false
 
-  private[hammurabi] def hasNext() = !values.isEmpty && !finished
+  private[hammurabi] def hasNext = values.nonEmpty && !finished
 
-  private[hammurabi] def +=[A] (t: Traversable[A]): Option[A] = {
+  private[hammurabi] def +=[A](t: Traversable[A]): Option[A] = {
     values = t.toList :: values
     finished = t.isEmpty
     if (finished) None else Some(t.head)
@@ -155,7 +157,7 @@ private class ValuesCombinator {
 
   private[hammurabi] def next[A](clazz: Class[A]): Option[A] = {
     finished = !valuesIterator.hasNext
-    if (finished) None else Some(valuesIterator.next.asInstanceOf[A])
+    if (finished) None else Some(valuesIterator.next().asInstanceOf[A])
   }
 
   private def cartesianProduct[A <: Any](l: List[List[A]]) = (l :\ List(List[A]())) {
