--- src/test/scala/hammurabi/IPerson.scala
@@ -1,3 +1,0 @@
-package hammurabi;
-
-case class IPerson(name: String, pos: Int = 0, color: String = null)
