+++ src/test/scala/hammurabi/ImmutablePerson.scala
@@ -1,0 +1,6 @@
+package hammurabi
+
+/**
+ * @author Mario Fusco
+ */
+case class ImmutablePerson(name: String, pos: Int = 0, color: String = null)
